initial test case for the '[...]' compound
